+++ libs/angular-modal-gallery/src/directives/size.directive.ts
@@ -23,7 +23,8 @@
  */
 
 import { Directive, ElementRef, Input, OnChanges, OnInit, Renderer2 } from '@angular/core';
-import { Size } from '../model/size.interface';
+
+import { Size } from '../model';
 
 /**
  * Directive to change the size of an element.